Handle paths with a '?' in them
This is actually quite tricky, as URI does not let you create these by
default, you need to sneak them in using URL.toUrl(). If you attempt to
use the URI constructor directly then a '?' is interpreted as the start
of the query string, and %3f is encoded to %253f, so there is no way to
actually pass in an ecoded question mark to the URI constructor.